+++ tests/test_historical_client.py
@@ -126,7 +126,8 @@ class TestHistoricalClient:
             ("schema", "definition"),
             ("stype_in", "native"),
             ("stype_out", "product_id"),
-            ("encoding", "dbz"),
+            ("encoding", "dbn"),
+            ("compression", "zstd"),
         ]
         assert sorted(call["headers"].keys()) == ["accept", "user-agent"]
         assert call["headers"]["accept"] == "application/json"